chatbot/requirements.txt
# Chatbot dependencies (separate from main project)
chainlit>=1.0.0
httpx>=0.25.0
openai-whisper>=20231117
chromadb>=0.4.0
transformers>=4.40.0
torch>=2.0.0
numpy>=1.24.0
pyyaml>=6.0.0
